Logic changed
